requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.18
langchainhub==0.1.21
chromadb==0.5.20
streamlit==1.40.2
python-dotenv==1.1.0
openai==1.78.1
tiktoken==0.9.0
PyMuPDF==1.25.5
python-pptx==1.0.2
pdf2image==1.17.0
pdfminer.six==20250506
pikepdf==9.7.0
